+++ modules/AWSPowerShell/AWSPowerShellCompleters.psm1
@@ -16135,7 +16135,7 @@ $EC2_Completers = {
         # Amazon.EC2.ImageAttributeName
         "Get-EC2ImageAttribute/Attribute"
         {
-            $v = "blockDeviceMapping","bootMode","description","kernel","lastLaunchedTime","launchPermission","productCodes","ramdisk","sriovNetSupport"
+            $v = "blockDeviceMapping","bootMode","description","kernel","lastLaunchedTime","launchPermission","productCodes","ramdisk","sriovNetSupport","tpmSupport","uefiData"
             break
         }
 
@@ -16520,6 +16520,13 @@ $EC2_Completers = {
             break
         }
 
+        # Amazon.EC2.TpmSupportValues
+        "Register-EC2Image/TpmSupport"
+        {
+            $v = "v2.0"
+            break
+        }
+
         # Amazon.EC2.TrafficDirection
         {
             ($_ -eq "Edit-EC2TrafficMirrorFilterRule/TrafficDirection") -Or
@@ -16728,6 +16735,7 @@ $EC2_map = @{
     "TargetCapacityUnitType"=@("Get-EC2SpotPlacementScore")
     "TargetEnvironment"=@("New-EC2InstanceExportTask")
     "Tenancy"=@("Add-EC2CapacityReservation","Edit-EC2InstancePlacement","New-EC2CapacityReservationFleet")
+    "TpmSupport"=@("Register-EC2Image")
     "TrafficDirection"=@("Edit-EC2TrafficMirrorFilterRule","New-EC2TrafficMirrorFilterRule")
     "TrafficType"=@("New-EC2FlowLog")
     "TransportProtocol"=@("New-EC2ClientVpnEndpoint")
@@ -17157,6 +17165,7 @@ $EC2_SelectMap = @{
                "Get-EC2GroupsForCapacityReservation",
                "Get-EC2HostReservationPurchasePreview",
                "Get-EC2InstanceTypesFromInstanceRequirement",
+               "Get-EC2InstanceUefiData",
                "Get-EC2IpamAddressHistory",
                "Get-EC2IpamPoolAllocation",
                "Get-EC2IpamPoolCidr",
@@ -17940,7 +17949,7 @@ $EKS_Completers = {
         # Amazon.EKS.AMITypes
         "New-EKSNodegroup/AmiType"
         {
-            $v = "AL2_ARM_64","AL2_x86_64","AL2_x86_64_GPU","BOTTLEROCKET_ARM_64","BOTTLEROCKET_x86_64","CUSTOM"
+            $v = "AL2_ARM_64","AL2_x86_64","AL2_x86_64_GPU","BOTTLEROCKET_ARM_64","BOTTLEROCKET_ARM_64_NVIDIA","BOTTLEROCKET_x86_64","BOTTLEROCKET_x86_64_NVIDIA","CUSTOM"
             break
         }