+++ modules/AWSPowerShell/Cmdlets/ComputeOptimizer/Basic/Remove-CORecommendationPreference-Cmdlet.cs
@@ -85,7 +85,7 @@ namespace Amazon.PowerShell.Cmdlets.CO
         /// <para>
         /// <para>The target resource type of the recommendation preference to delete.</para><para>The <code>Ec2Instance</code> option encompasses standalone instances and instances
         /// that are part of Auto Scaling groups. The <code>AutoScalingGroup</code> option encompasses
-        /// only instances that are part of an Auto Scaling group.</para>
+        /// only instances that are part of an Auto Scaling group.</para><note><para>The valid values for this parameter are <code>Ec2Instance</code> and <code>AutoScalingGroup</code>.</para></note>
         /// </para>
         /// </summary>
         #if !MODULAR