+++ modules/AWSPowerShell/Cmdlets/EKS/Basic/Get-EKSIdentityProviderConfig-Cmdlet.cs
@@ -77,7 +77,7 @@ namespace Amazon.PowerShell.Cmdlets.EKS
         #region Parameter IdentityProviderConfig_Type
         /// <summary>
         /// <para>
-        /// <para>The type of the identity provider configuration.</para>
+        /// <para>The type of the identity provider configuration. The only type available is <code>oidc</code>.</para>
         /// </para>
         /// </summary>
         #if !MODULAR