+++ CHANGELOG.md
@@ -1,4 +1,12 @@
-﻿### 4.1.81 (2022-05-09 20:18Z)
+﻿### 4.1.82 (2022-05-10 20:12Z)
+  * AWS Tools for PowerShell now use AWS .NET SDK 3.7.271.0 and leverage its new features and improvements. Please find a description of the changes at https://github.com/aws/aws-sdk-net/blob/master/SDK.CHANGELOG.md.
+  * Amazon Elastic Compute Cloud (EC2)
+    * Added cmdlet Get-EC2InstanceUefiData leveraging the GetInstanceUefiData service API.
+    * Modified cmdlet Register-EC2Image: added parameters TpmSupport and UefiData.
+  * Amazon Elastic MapReduce
+    * Modified cmdlet Start-EMRJobFlow: added parameter OSReleaseLabel.
+
+### 4.1.81 (2022-05-09 20:18Z)
   * AWS Tools for PowerShell now use AWS .NET SDK 3.7.270.0 and leverage its new features and improvements. Please find a description of the changes at https://github.com/aws/aws-sdk-net/blob/master/SDK.CHANGELOG.md.
 
 ### 4.1.80 (2022-05-06 20:22Z)